+++ src/aim/cftemplates/__init__.py
@@ -18,7 +18,6 @@ from aim.cftemplates.nat_gateway import NATGateway
 from aim.cftemplates.kms import KMS
 from aim.cftemplates.cw_alarms import CWAlarms
 from aim.cftemplates.lambda_function import Lambda
-from aim.cftemplates.lambda_permission import LambdaPermission
 from aim.cftemplates.account import Account
 from aim.cftemplates.eventsrule import EventsRule
 from aim.cftemplates.snstopics import SNSTopics