+++ src/aim/core/yaml.py
@@ -8,7 +8,6 @@ class Ref:
     def __init__(self, value, style=None):
         self.value = value
         self.style = style
-        breakpoint()
 
     @classmethod
     def to_yaml(cls, representer, node):
@@ -38,12 +37,13 @@ class Sub:
     def from_yaml(cls, constructor, node):
         return cls(node.value, node.style)
 
+
 class YAML(ruamel.yaml.YAML):
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
-#        ruamel.yaml.add_constructor(u'!Ref', self.tag_constructor, constructor=ruamel.yaml.SafeConstructor)
-#        ruamel.yaml.add_constructor(u'!Sub', self.tag_constructor, constructor=ruamel.yaml.SafeConstructor)
+        ruamel.yaml.add_constructor('!Ref', self.tag_constructor, constructor=ruamel.yaml.SafeConstructor)
+        ruamel.yaml.add_constructor('!Sub', self.tag_constructor, constructor=ruamel.yaml.SafeConstructor)
 
     def tag_constructor(self, node, arg2):
         return arg2.tag+ ' '+arg2.value
@@ -56,3 +56,10 @@ class YAML(ruamel.yaml.YAML):
         ruamel.yaml.YAML.dump(self, data, stream, **kw)
         if dumps:
             return stream.getvalue()
+
+def read_yaml_file(path):
+    yaml = YAML(typ="safe", pure=True)
+    yaml.default_flow_sytle = False
+    with open(path, 'r') as stream:
+        data = yaml.load(stream)
+    return data